add endpoint for adding an activity log
Note: the endpoint does not check if you have provided a valid user ID, valid description, etc. In particular, if provided with an invalid user ID, then there may a foreign key constraint violated and the log will not be added (in addition an error message will print out). The endpoint returns true if a log was successfully added and false otherwise.